refactor: split expression inline scan helpers
